CAMBIOS

+++ src/main/java/com/peruvianfarma/appweb/model/Producto.java
@@ -20,16 +20,16 @@ import lombok.*;
 @Builder
 @Entity
 @Table(name = "t_product")
-public class Producto {
+public class Producto{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String descripcion;
     private BigDecimal precio;
     @Temporal(TemporalType.DATE)
     private Date dueDate;
     private String imageFileName;
     private String status;
-    
+   
 }
